AND-3631: If device time is in past and a message is received on the device then the time of message received is displayed in further past of the device current time.
RCA: The check to prevent receiving messages from future in convMessage is taken care by Utils.applyServerTimeOffset already. Having this would readjust the date twice and cause the issue
This check in convMessage was added first(Jan 2012) and when we added code for serverTimeOffset later(Jun 2013) we probably forgot to remove convMessage code.